Add State::apply_proof and notifications module
